requirements.txt
flask==3.0.0
elevenlabs==0.2.27
python-dotenv==1.0.0
